+++ packages/mongo/test/index.js
@@ -14,10 +14,8 @@ test('Collection option merges into default options', t => {
   const store = new KeyvMongo({ collection: 'foo' })
   t.deepEqual(store.options, {
     url: 'mongodb://127.0.0.1:27017',
-    mongoOptions: {
-      useNewUrlParser: true,
-      useUnifiedTopology: true
-    },
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
     collection: 'foo',
     emitErrors: true
   })
@@ -27,10 +25,8 @@ test('Collection option merges into default options if URL is passed', t => {
   const store = new KeyvMongo(mongoURL, { collection: 'foo' })
   t.deepEqual(store.options, {
     url: mongoURL,
-    mongoOptions: {
-      useNewUrlParser: true,
-      useUnifiedTopology: true
-    },
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
     collection: 'foo',
     emitErrors: true
   })