+++ packages/sqlite/test/sqlite.test-d.ts
@@ -1,0 +1,12 @@
+import Keyv = require('@keyvhq/core');
+import KeyvSqlite = require('..');
+
+new Keyv({ store: new KeyvSqlite({ uri: 'sqlite://path/to/database.sqlite', table: 'cache' }) });
+
+new KeyvSqlite({ uri: 'sqlite://path/to/database.sqlite' });
+new KeyvSqlite({ busyTimeout: 10000 });
+new KeyvSqlite({ table: 'cache' });
+new KeyvSqlite({ keySize: 100 });
+
+const sqlite = new KeyvSqlite({ uri: 'sqlite://path/to/database.sqlite' });
+new Keyv({ store: sqlite });