+++ packages/redis/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1](https://github.com/microlinkhq/keyv/compare/v1.5.0...v1.5.1) (2021-09-16)
+
+**Note:** Version bump only for package @keyvhq/redis
+
+
+
+
+
 # [1.4.0](https://github.com/microlinkhq/keyv/compare/v1.3.0...v1.4.0) (2021-09-12)